# openapi/customer-phone-api_1.0.0.yml
openapi: 3.0.0
info:
  title: Customer Phone Number Management API
  version: 1.0.0
  description: Belong | Demonstration API | Manages phone numbers associated with customers.

paths:
  /phone-numbers:
    get:
      summary: Get all phone numbers with pagination.
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          example: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          example: 10
      responses:
        '200':
          description: A list of phone numbers
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/PhoneNumber'
                  page:
                    type: integer
                  size:
                    type: integer
                  total_pages:
                    type: integer
                  total_results:
                    type: integer
              example:
                page: 1
                size: 10
                total_pages: 1
                total_results: 3
                results:
                  - id: 1
                    number: "0412334423"
                    active: true
                    customer_id: 1
                  - id: 2
                    number: "0499883788"
                    active: false
                    customer_id: 1
                  - id: 3
                    number: "0490227733"
                    active: false
                    customer_id: 2
        '400':
          description: Input validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_page_size:
                  value:
                    request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                    message: "Page size should be grater than or equal to one."
                invalid_page_number:
                  value:
                    request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                    message: "Page number should be grater than or equal to zero."
        '500':
          description: Unexpected failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Unreachable database."

  /phone-numbers/customers/{customer_id}:
    get:
      summary: Get all phone numbers for a specific customer.
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
          example: 1
      responses:
        '200':
          description: A list of phone numbers for the given customer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhoneNumber'
              examples:
                customer_found:
                  value:
                    - id: 1
                      number: "0412334423"
                      active: true
                      customer_id: 1
                    - id: 2
                      number: "0499883788"
                      active: true
                      customer_id: 1
                customer_not_found:
                  value: [ ]
        '400':
          description: Input validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Customer ID must be a numeric."
        '500':
          description: Unexpected failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Unreachable database."

  /phone-numbers/{phone_number_id}/activation:
    patch:
      summary: Activate a phone number.
      parameters:
        - name: phone_number_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Phone number activation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneNumber'
              example:
                id: 1
                number: "0412334423"
                active: true
                customer_id: 1
        '400':
          description: Input validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Phone number ID format invalid."
        '404':
          description: Input validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Phone number not found."
        '409':
          description: Cannot update phone number status, already in desired state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Cannot update phone number status, already in desired state"
        '500':
          description: Unexpected failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Unreachable database."

  /phone-numbers/{phone_number_id}/deactivation:
    patch:
      summary: Deactivate a phone number. At least one phone number must be active.
      parameters:
        - name: phone_number_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Phone number deactivation successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneNumber'
              example:
                id: 1
                number: "0412334423"
                active: true
                customer_id: 1
        '400':
          description: Input validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Phone number ID format invalid."
        '404':
          description: Input validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Phone number not found."
        '409':
          description: Cannot update phone number status, already in desired state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Cannot update phone number status, already in desired state"
        '500':
          description: Unexpected failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                request_id: "558c083d-7acc-46c6-9ab0-1616c8c7938a"
                message: "Unreachable database."

components:
  schemas:
    PhoneNumber:
      type: object
      properties:
        id:
          type: integer
        number:
          type: string
        active:
          type: boolean
        customer_id:
          type: integer
      example:
        id: "1"
        number: "+1234567890"
        active: true
        customer_id: "c123"
    ErrorResponse:
      type: object
      properties:
        request_id:
          type: string
        message:
          type: string